requirements.txt
absl-py==0.12.0
alembic==1.4.1
altair==4.1.0
appdirs==1.4.4
arch==4.19
argh==0.26.2
argon2-cffi==20.1.0
asgiref==3.3.4
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
backports.weakref==1.0.post1
backtrader==1.9.76.123
black==19.10b0
bleach==3.3.0
blinker==1.4
blis==0.7.4
Boruta==0.3
cachetools==4.2.2
catalogue==1.0.0
catboost==0.25.1
category-encoders==2.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
colorlover==0.3.0
cryptography==3.4.7
cufflinks==0.17.3
cv2-tools==2.4.0
cycler==0.10.0
cymem==2.0.5
cython==0.29.23
databricks-cli==0.14.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.2.4
docker==5.0.0
docutils==0.16
entrypoints==0.3
Flask==2.0.1
flatbuffers==1.12
funcy==1.16
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.16
greenlet==1.1.0
grpcio==1.34.1
h11==0.12.0
h5py==3.1.0
HeapDict==1.0.1
helpdev==0.7.1
htmlmin==0.1.12
idna==2.10
image==1.5.33
ImageHash==4.2.0
imageio==2.9.0
imbalanced-learn==0.7.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab==2.2.6
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupyterlab-widgets==1.0.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kmodes==0.11.0
lightgbm==3.2.1
llvmlite==0.36.0
locket==0.2.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
mlflow==1.17.0
mlxtend==0.18.0
multimethod==1.4
murmurhash==1.0.5
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nltk==3.6.2
notebook==6.4.0
numba==0.53.1
numexpr==2.7.3
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.2.52
opencv-python==4.5.2.52
opt-einsum==3.3.0
outcome==1.1.0
p5py==1.0.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.7.0
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.1
pep517==0.10.0
phik==0.11.2
pickleshare==0.7.5
Pillow==8.2.0
pkginfo==1.7.0
plac==1.1.3
plotly==4.14.3
pluggy==0.13.1
ply==3.11
preshed==3.0.5
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
prompt-toolkit==3.0.19
property-cached==1.6.4
protobuf==3.15.8
pyarrow==4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.9.0
pyLDAvis==3.2.2
pynndescent==0.5.2
pyod==0.8.8
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-constraint==1.4.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
QDarkStyle==2.8.1
qgrid==1.3.1
QtPy==1.9.0
querystring-parser==1.2.4
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
ruamel.yaml==0.15.87
scikit-image==0.18.1
scikit-learn==0.24.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.1
selenium==4.0.0a7
Send2Trash==1.5.0
six==1.15.0
smart-open==5.1.0
smmap==4.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
spacy==2.3.5
SQLAlchemy==1.4.16
sqlparse==0.4.1
srsly==1.0.5
statsmodels==0.12.2
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.5.0
textblob==0.15.3
thinc==7.4.5
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
trio==0.18.0
trio-websocket==0.9.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
umap-learn==0.5.1
urllib3==1.26.5
waitress==2.0.0
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
wsproto==1.0.0
xgboost==1.4.2
xmltodict==0.12.0
yellowbrick==1.3.post1
zict==2.0.0
